@@ -1,28 +1,28 @@
 name: example_image_picker_plus
 description: A new custom_gallery_display project.
 
-publish_to: 'none' 
+publish_to: 'none'
 
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.5 <4.0.0'
+  sdk: ^3.11.0
+
 
 dependencies:
   flutter:
     sdk: flutter
   image_picker_plus:
     path: ../
-  video_player: ^2.7.0
+  video_player: ^2.9.2
 
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.8
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.0
+  flutter_lints: ^6.0.0
 
 flutter:
-
   uses-material-design: true
